+++ src/modules/address/index.ts
@@ -3,6 +3,7 @@ import { TIME_ZONE } from "./constants/time-zone";
 import { CARDINAL_DIRECTIONS } from "./constants/cardinal-directions";
 import { ORDINAL } from "./constants/ordinal";
 import { ChacaUtils } from "../../core/utils";
+import { DatatypeModule } from "../datatype";
 
 export type ZipCodeProps = {
   format?: string;
@@ -19,8 +20,23 @@ export type CountryProps = {
     | "Antartica";
 };
 
+export type LatitudeProps = {
+  max?: number;
+  min?: number;
+  precision?: number;
+};
+
+export type LongitudProps = {
+  max?: number;
+  min?: number;
+  precision?: number;
+};
+
 export class AddressModule {
-  constructor(private readonly utils: ChacaUtils) {}
+  constructor(
+    private readonly utils: ChacaUtils,
+    private readonly datatypeModule: DatatypeModule,
+  ) {}
 
   readonly constants = {
     timeZones: TIME_ZONE,
@@ -98,5 +114,51 @@ export class AddressModule {
    */
   ordinalDirection(): string {
     return this.utils.oneOfArray(ORDINAL);
+  }
+
+  /**
+   * Generates a random latitude.
+   *
+   * @param options.max The upper bound for the latitude to generate. Defaults to `90`.
+   * @param options.min The lower bound for the latitude to generate. Defaults to `-90`.
+   * @param options.precision The number of decimal points of precision for the latitude. Defaults to `4`.
+   *
+   * @example
+   * modules.address.latitude() // -30.9501
+   * modules.address.latitude({ max: 10 }) // 5.7225
+   * modules.address.latitude({ max: 10, min: -10 }) // -9.6273
+   * modules.address.latitude({ max: 10, min: -10, precision: 5 }) // 2.68452
+   */
+  latitude(options: LatitudeProps = {}): number {
+    const { max = 90, min = -90, precision = 4 } = options;
+
+    return this.datatypeModule.float({
+      min: min,
+      max: max,
+      precision: precision,
+    });
+  }
+
+  /**
+   * Generates a random longitude.
+   *
+   * @param options.max The upper bound for the longitude to generate. Defaults to `180`.
+   * @param options.min The lower bound for the longitude to generate. Defaults to `-180`.
+   * @param options.precision The number of decimal points of precision for the longitude. Defaults to `4`.
+   *
+   * @example
+   * modules.address.longitude() // -30.9501
+   * modules.address.longitude({ max: 10 }) // 5.7225
+   * modules.address.longitude({ max: 10, min: -10 }) // -9.6273
+   * modules.address.longitude({ max: 10, min: -10, precision: 5 }) // 2.68452
+   */
+  longitude(options: LongitudProps = {}): number {
+    const { max = 180, min = -180, precision = 4 } = options;
+
+    return this.datatypeModule.float({
+      max: max,
+      min: min,
+      precision: precision,
+    });
   }
 }