+++ test/library/example-case/core/definitions/couchbase/index.ts
@@ -10,5 +10,5 @@ export const COUCHBASE_DATASET = chaca.dataset([
   { name: "Airport", documents: 10, schema: AIRPORT_SCHEMA },
   { name: "Hotel", documents: 60, schema: HOTEL_SCHEMA },
   { name: "Route", documents: 200, schema: ROUTE },
-  { name: "User", documents: 500, schema: USER },
+  { name: "User", documents: 800, schema: USER },
 ]);