+++ test/library/field-types/nested-object/nested-object.test.ts
@@ -2,10 +2,9 @@ import { chaca, modules } from "../../../../src";
 import { describe, expect, it } from "vitest";
 
 describe("Nested object field", () => {
-  it("Should return an object with user field as an object", () => {
+  it("define a user nested schema. should return an object with user field as an object", () => {
     const schema = chaca.schema({
       id: () => modules.id.uuid(),
-      image: () => modules.image.people(),
       user: chaca.schema({
         username: () => modules.internet.username(),
         image: () => modules.image.fashion(),
@@ -18,7 +17,7 @@ describe("Nested object field", () => {
     expect(doc.user).toHaveProperty("image");
   });
 
-  it("Should return an object with a user field with the image field as array of string", () => {
+  it("should return an object with a user field with the image field as array of string", () => {
     const schema = chaca.schema({
       user: chaca.schema({
         username: () => modules.internet.username(),
@@ -31,7 +30,7 @@ describe("Nested object field", () => {
     expect(doc.user.images.length).toBe(10);
   });
 
-  it("Should return an object with a user field as an array of objects with image and username property", () => {
+  it("define user schema, isArray = 10. should return an object with a user field as an array of objects with image and username property", () => {
     const schema = chaca.schema({
       user: {
         type: chaca.schema({
@@ -45,5 +44,10 @@ describe("Nested object field", () => {
     const doc = schema.object();
 
     expect(doc.user.length).toBe(20);
+
+    for (const user of doc.user) {
+      expect(user).toHaveProperty("username");
+      expect(user).toHaveProperty("image");
+    }
   });
 });